Extract generic table diff into shared spf::diff module
Move TableDiffResult and table_diff() from ospf/inst.rs and isis/inst.rs
into spf/diff.rs to eliminate code duplication between protocols.
