+++ zebra-rs/src/ospf/inst.rs
@@ -1368,80 +1368,14 @@ fn perform_spf_calculation(top: &mut Ospf, area_id: Ipv4Addr) {
     }
 }
 
-/// Generic result for table diff operations
-#[derive(Debug)]
-pub struct TableDiffResult<'a, K, V> {
-    pub only_curr: Vec<(&'a K, &'a V)>,
-    pub only_next: Vec<(&'a K, &'a V)>,
-    pub different: Vec<(&'a K, &'a V, &'a V)>,
-    pub identical: Vec<(&'a K, &'a V)>,
-}
-
-pub type DiffResult<'a> = TableDiffResult<'a, Ipv4Net, SpfRoute>;
+pub type DiffResult<'a> = spf::TableDiffResult<'a, Ipv4Net, SpfRoute>;
 
 /// Convenience function for SPF route diffs (backward compatibility)
 pub fn diff<'a>(
     curr: &'a PrefixMap<Ipv4Net, SpfRoute>,
     next: &'a PrefixMap<Ipv4Net, SpfRoute>,
 ) -> DiffResult<'a> {
-    table_diff_impl(curr.iter(), next.iter())
-}
-
-/// Generic table diff implementation
-fn table_diff_impl<'a, K, V, I>(curr_iter: I, next_iter: I) -> TableDiffResult<'a, K, V>
-where
-    K: Ord,
-    V: PartialEq,
-    I: Iterator<Item = (&'a K, &'a V)>,
-{
-    let mut res = TableDiffResult {
-        only_curr: vec![],
-        only_next: vec![],
-        different: vec![],
-        identical: vec![],
-    };
-
-    let mut curr_iter = curr_iter.peekable();
-    let mut next_iter = next_iter.peekable();
-
-    while let (Some(&(curr_key, curr_value)), Some(&(next_key, next_value))) =
-        (curr_iter.peek(), next_iter.peek())
-    {
-        match curr_key.cmp(next_key) {
-            std::cmp::Ordering::Less => {
-                // curr_key is only in curr
-                res.only_curr.push((curr_key, curr_value));
-                curr_iter.next();
-            }
-            std::cmp::Ordering::Greater => {
-                // next_key is only in next
-                res.only_next.push((next_key, next_value));
-                next_iter.next();
-            }
-            std::cmp::Ordering::Equal => {
-                // keys are equal; compare values
-                if curr_value == next_value {
-                    res.identical.push((curr_key, curr_value));
-                } else {
-                    res.different.push((curr_key, curr_value, next_value));
-                }
-                curr_iter.next();
-                next_iter.next();
-            }
-        }
-    }
-
-    // Deal with the rest of curr
-    for (key, value) in curr_iter {
-        res.only_curr.push((key, value));
-    }
-
-    // Deal with the rest of next
-    for (key, value) in next_iter {
-        res.only_next.push((key, value));
-    }
-
-    res
+    spf::table_diff(curr.iter(), next.iter())
 }
 
 fn nhop_to_nexthop_uni(key: &Ipv4Addr, route: &SpfRoute, value: &SpfNexthop) -> rib::NexthopUni {
