+++ zebra-rs/src/spf/mod.rs
@@ -1,2 +1,5 @@
 pub mod calc;
 pub use calc::*;
+
+pub mod diff;
+pub use diff::*;
